test(cli): split config tests
